+++ checkers_bot_tournament/bots/greedycat.py
