+++ checkers_bot_tournament/bots/scaredycat.py
